Mark insert and delete INLINABLE.
So callsite can specialize or inline them as it wills. Some noticable
imprevement on benchmarks:

```
insert: 16ms -> 11ms
delete: 26ms -> 16ms

```

Also added the splitCuckooFilter function so benchmark can start from a
pre-filled CF. Could have made it that either it only returns a new CF,
or returns a pair whose fst is the original CF (with the same ByteArray)
but with the split gen.. but those leave too much opportunity for
accidental misuse.